app/requirements.txt
psycopg2-binary==2.9.10
fastapi==0.116.1
pydantic==2.11.7
dotenv==0.9.9
uvicorn>=0.35.0
